Adds social media to users and bands
Fields for twitter, facebook, and myspace added to both users and bands. If band has a twitter id entered, then use JS to pull in latest 3 tweets. Proof of concept/starting point. Could use to pull in number of followers and the like? Facebook id will be more interesting - need to be able to import from facebook, which doesn't require storing the id, but a link to their facebook id isn't a bad idea regardless.